+++ pkg/combustion/custom.go
@@ -41,30 +41,52 @@ func configureCustomFiles(ctx *image.Context) ([]string, error) {
 
 func handleCustomFiles(ctx *image.Context) error {
 	fullFilesDir := generateComponentPath(ctx, filepath.Join(customDir, customFilesDir))
-	_, err := copyCustomFiles(fullFilesDir, ctx.CombustionDir, nil)
+	err := copyCustomFiles(fullFilesDir, ctx.CombustionDir)
 	return err
 }
 
 func handleCustomScripts(ctx *image.Context) ([]string, error) {
 	fullScriptsDir := generateComponentPath(ctx, filepath.Join(customDir, customScriptsDir))
 	executablePerms := fileio.ExecutablePerms
-	scripts, err := copyCustomFiles(fullScriptsDir, ctx.CombustionDir, &executablePerms)
+	scripts, err := copyCustomScripts(fullScriptsDir, ctx.CombustionDir, &executablePerms)
 	return scripts, err
 }
 
-func copyCustomFiles(fromDir, toDir string, filePermissions *os.FileMode) ([]string, error) {
+func copyCustomFiles(fromDir, toDir string) error {
+	if _, err := os.Stat(fromDir); os.IsNotExist(err) {
+		return nil
+	}
+
+	dirEntries, err := os.ReadDir(fromDir)
+	if err != nil {
+		return fmt.Errorf("reading the custom files directory at %s: %w", fromDir, err)
+	}
+
+	// If the directory exists but there's nothing in it, consider it an error case
+	if len(dirEntries) == 0 {
+		return fmt.Errorf("no files found in directory %s", fromDir)
+	}
+
+	if err = fileio.CopyFiles(fromDir, toDir, "", true); err != nil {
+		return fmt.Errorf("copying custom files and directories: %w", err)
+	}
+
+	return nil
+}
+
+func copyCustomScripts(fromDir, toDir string, filePermissions *os.FileMode) ([]string, error) {
 	if _, err := os.Stat(fromDir); os.IsNotExist(err) {
 		return nil, nil
 	}
 
 	dirEntries, err := os.ReadDir(fromDir)
 	if err != nil {
-		return nil, fmt.Errorf("reading the custom directory at %s: %w", fromDir, err)
+		return nil, fmt.Errorf("reading the custom scripts directory at %s: %w", fromDir, err)
 	}
 
 	// If the directory exists but there's nothing in it, consider it an error case
 	if len(dirEntries) == 0 {
-		return nil, fmt.Errorf("no files found in directory %s", fromDir)
+		return nil, fmt.Errorf("no scripts found in directory %s", fromDir)
 	}
 
 	var copiedFiles []string
@@ -73,19 +95,8 @@ func copyCustomFiles(fromDir, toDir string, filePermissions *os.FileMode) ([]str
 		copyMe := filepath.Join(fromDir, entry.Name())
 		copyTo := filepath.Join(toDir, entry.Name())
 
-		var mode os.FileMode
-		if filePermissions == nil {
-			info, infoErr := entry.Info()
-			if infoErr != nil {
-				return nil, fmt.Errorf("reading file info: %w", infoErr)
-			}
-			mode = info.Mode()
-		} else {
-			mode = *filePermissions
-		}
-
-		if err = fileio.CopyFile(copyMe, copyTo, mode); err != nil {
-			return nil, fmt.Errorf("copying file to %s: %w", copyTo, err)
+		if err = fileio.CopyFile(copyMe, copyTo, *filePermissions); err != nil {
+			return nil, fmt.Errorf("copying script to %s: %w", copyTo, err)
 		}
 
 		copiedFiles = append(copiedFiles, entry.Name())